refactor & directory changes
