+++ docs/FLIGHT_GUIDE.md
@@ -11,27 +11,34 @@
 ### 2. Start Ground Station
 ```bash
 cd gds
-python advanced_web_gui.py
+
+# Activate venv (macOS/Linux)
+source venv/bin/activate
+
+# Activate venv (Windows PowerShell)
+.\venv\Scripts\Activate.ps1
+
+python slugsight_gds.py
 ```
 
 **Expected Output:**
 ```
-INFO - Data logging to: gds/flight_data/slugsight_20251109_120000.csv
+INFO - Data logging to: .../gds/flight_data/slugsight_20251109_120000.csv
 INFO - Found GCS Receiver on port: /dev/cu.usbserial-XXX
-INFO - Open this URL in your browser: http://127.0.0.1:5000
+INFO - Open this URL in your browser: [http://127.0.0.1:8080](http://127.0.0.1:8080)
 ```
 
 ### 3. Open Web Dashboard
-1. Open browser to: http://127.0.0.1:5000
+1. Open browser to: [http://127.0.0.1:8080](http://127.0.0.1:8080)
 2. Verify "Connected" status (green indicator)
 3. Check data is updating (numbers changing)
 
 ### 4. Pre-Flight Checks
-- [ ] **Battery**: VBat > 3.7V (preferably > 4.0V)
+- [ ] **Battery**: VBat > 3.7V (Appears immediately on boot)
 - [ ] **Signal**: RSSI > -100 dBm (closer to 0 is better)
-- [ ] **GPS**: Wait for GPS fix (may take 30-60 seconds)
+- [ ] **GPS**: Wait for GPS fix (may take 30-60 seconds). Fix status will update from '0' to '1' when locked.
 - [ ] **Sensors**: Altitude reading reasonable
-- [ ] **Data Logging**: Note CSV filename from terminal
+- [ ] **Data Logging**: Confirm Ground Station is recording (see terminal). **Onboard SD logging** will start automatically once GPS fix is acquired.
 
 ## During Flight
 
@@ -52,17 +59,18 @@ INFO - Open this URL in your browser: http://127.0.0.1:5000
 ### 1. Verify Data Saved
 ```bash
 ls -lh gds/flight_data/
+# On Windows: dir gds\flight_data\
 ```
 
 Should show your flight CSV file with size > 0 bytes
 
 ### 2. Quick Data Check
 ```bash
-# View first 10 rows
+# View first 10 rows (macOS/Linux)
 head -10 gds/flight_data/slugsight_YYYYMMDD_HHMMSS.csv
 
-# Count packets received
-wc -l gds/flight_data/slugsight_YYYYMMDD_HHMMSS.csv
+# View first 10 rows (Windows PowerShell)
+Get-Content gds/flight_data/slugsight_YYYYMMDD_HHMMSS.csv -Head 10
 ```
 
 ### 3. Shutdown Ground Station
@@ -70,54 +78,24 @@ wc -l gds/flight_data/slugsight_YYYYMMDD_HHMMSS.csv
 - Wait for "Ground station shutdown complete" message
 - CSV file is automatically closed and saved
 
-## Quick Analysis
-
-### Find Apogee
-```python
-import pandas as pd
-df = pd.read_csv('gds/flight_data/YOUR_FILE.csv')
-apogee = df['Altitude'].max()
-print(f"Apogee: {apogee:.1f} meters")
-```
-
-### Plot Flight Profile
-```python
-import pandas as pd
-import matplotlib.pyplot as plt
-
-df = pd.read_csv('gds/flight_data/YOUR_FILE.csv')
-time = df.index * 0.1  # 10 Hz data
-
-fig, (ax1, ax2) = plt.subplots(2, 1, figsize=(12, 8))
-
-# Altitude
-ax1.plot(time, df['Altitude'])
-ax1.set_ylabel('Altitude (m)')
-ax1.set_title('Flight Profile')
-ax1.grid(True)
-
-# Velocity
-ax2.plot(time, df['Velocity'])
-ax2.set_xlabel('Time (s)')
-ax2.set_ylabel('Velocity (m/s)')
-ax2.grid(True)
-
-plt.tight_layout()
-plt.savefig('flight_profile.png')
-plt.show()
-```
-
 ## Troubleshooting
 
 ### "Could not find GCS Receiver"
+**1. macOS/Linux:**
 ```bash
-# List USB devices (macOS)
 ls /dev/cu.*
-
-# List USB devices (Linux)
 ls /dev/ttyUSB* /dev/ttyACM*
 ```
-Update `ARDUINO_VID_PIDS` in `advanced_web_gui.py` if needed
+
+**2. Windows:**
+- Open Device Manager > Ports (COM & LPT)
+- Look for "USB Serial Device" or "Arduino"
+- Or run this Python command to list all ports:
+```bash
+python -m serial.tools.list_ports
+```
+
+Update `ARDUINO_VID_PIDS` in `gds/slugsight_gds.py` if needed.
 
 ### No Data Showing
 1. Check receiver serial monitor in Arduino IDE
@@ -125,24 +103,10 @@ Update `ARDUINO_VID_PIDS` in `advanced_web_gui.py` if needed
 3. Check transmitter battery voltage
 4. Verify transmitter is sending (LED should blink)
 
-### Dashboard Shows Old Data
-- Refresh browser page
-- Check WebSocket connection (green status)
-- Restart ground station
-
 ### CSV File Empty or Missing
 - Check terminal for error messages
 - Verify write permissions on `gds/flight_data/` directory
 - Check disk space
-
-## Data Rate Reference
-
-At 10 Hz transmission rate:
-- **10 packets/second**
-- **600 packets/minute**
-- **1-minute flight** ≈ 600 rows
-- **5-minute flight** ≈ 3,000 rows
-- **CSV file size** ≈ 150-200 KB per minute
 
 ## Emergency Procedures
 
@@ -170,22 +134,3 @@ mkdir -p ~/rocket_backups
 # After flight, immediately backup data
 cp gds/flight_data/*.csv ~/rocket_backups/
 ```
-
-## Contact Info
-
-For issues during flight prep:
-- Check `gds/README.md` for detailed troubleshooting
-- Check `README.md` for system overview
-- Review `CHANGES.md` for recent updates
-
-## Success Checklist
-
-✅ Ground station running and connected  
-✅ Web dashboard showing live data  
-✅ CSV filename noted for post-flight  
-✅ GPS has fix (if needed)  
-✅ Battery voltage good (> 3.7V)  
-✅ RSSI acceptable (> -100 dBm)  
-✅ All readouts changing/updating  
-
-**You're ready to launch! 🚀**
