+++ gds/requirements.txt
@@ -1,3 +1,5 @@
-flask>=2.3.0
-flask-sock>=0.6.0
-pyserial>=3.5
+Flask==2.3.2
+Flask-Sock==0.5.0
+pyserial==3.5
+Werkzeug==2.3.7
+pytz
