--- gds/test_flight_data/test_20251109_161318.csv
@@ -1,6 +1,0 @@
-timestamp,packet_count,Pitch,Roll,Yaw,Altitude,Velocity,Accel X,Accel Y,Accel Z,Pressure Pa,IMU Temp C,GPS Fix,GPS Sats,GPS Lat,GPS Lon,GPS Alt m,GPS Speed m/s,VBat,RSSI
-2025-11-09T16:13:18.108422,0,0.0,0.0,0.0,0.0,0.0,0.0,0.0,9.81,101325.0,20.0,0,0,0.0,0.0,0.0,0.0,4.1,-100.0
-2025-11-09T16:13:18.108508,1,2.5,-1.2,15.3,25.5,5.2,0.3,0.1,9.9,101200.0,21.2,0,3,37.1,-122.3,28.0,3.5,4.05,-98.0
-2025-11-09T16:13:18.108538,2,8.3,-3.5,45.8,125.5,25.8,1.2,0.5,11.2,98500.0,23.5,1,7,37.123456,-122.345678,130.5,28.5,3.95,-92.0
-2025-11-09T16:13:18.108562,3,15.2,2.8,78.9,285.3,45.2,2.8,1.2,13.5,92000.0,25.8,1,9,37.145678,-122.36789,290.0,55.3,3.88,-85.0
-2025-11-09T16:13:18.108583,4,12.5,5.1,92.5,425.8,38.5,2.1,0.9,12.8,88500.0,27.2,1,10,37.16789,-122.389012,430.2,48.7,3.82,-83.0
